Make the methods of the IRC object thread-safe.
How it's done: We equip each IRC instance with a reentrant thread lock.  Then, method bodies that modify mutable shared state (the connection list, the handlers list, and the delayed-events list) are guarded by the mutex using the Python `with` statement, which implicitly invokes the lock acquire()/release() methods at its scope bundaries.

The body of the process_once() function is also guarded, so any event handlers that it fires also won't be able to asynchronously mess with the shared state.

These changes have been tested in irker.