Add a connection factory for the async client
Related to #146.

Using a factory lets users easily handle needed features. This doesn't fully
solve #146 at all, but allows for quick workarounds.

The default factory accepts random keywords during initialization, that will be
given to loop.create_connection. The parameters are different from Factory, but
they more or less cover the same features: IPv6 and SSL are now supported
out-of-the-box for example.

This change shouldn't break any existing code using the library. It can be
greatly improved (by improving constistency for example), but it's a start.